Improvements to block prefab
Copy the transmitter script into the receiver script. This allows us to have a single location to control whether we are emitting or receiving.
Changes to the ReceiverEmitterBlockController to use the changes to ReceiverTrigger- breaks use of old EmitterTrigger.
Update indicator materials as the trigger type changes for realtime feedback.
Add a MaterialManager so we can set materials in a single location.